Add files via upload
This folder contains various Eviews (statistical software package) codes where midas (mixed data sampling) models are applied to macroeconomic and financial variables. It also comprises in- and out-of-sample analyses for a specific variable 'commercial paper to treasury spread' and connections between the forecasting performance of volatility models with business cycle.